restructure packaging
