+++ spring-web-modules/spring-rest-http-3/src/main/java/com/baeldung/bulkandbatchapi/exception/BatchCreateException.java
@@ -1,4 +1,4 @@
-package com.baeldung.bulkandbatchapi;
+package com.baeldung.bulkandbatchapi.exception;
 
 public class BatchCreateException extends RuntimeException {
     public BatchCreateException(String exceptionMsg) {
