+++ spring-web-modules/spring-rest-http-3/src/main/java/com/baeldung/bulkandbatchapi/request/BatchRequest.java
@@ -1,4 +1,4 @@
-package com.baeldung.bulkandbatchapi;
+package com.baeldung.bulkandbatchapi.request;
 
 import jakarta.validation.constraints.NotNull;
