+++ spring-web-modules/spring-rest-http-3/src/main/java/com/baeldung/bulkandbatchapi/request/Customer.java
@@ -1,4 +1,4 @@
-package com.baeldung.bulkandbatchapi.customer;
+package com.baeldung.bulkandbatchapi.request;
 
 import jakarta.validation.constraints.NotBlank;
