+++ spring-web-modules/spring-rest-http-3/src/main/java/com/baeldung/bulkandbatchapi/request/CustomerBatchRequest.java
@@ -1,6 +1,5 @@
-package com.baeldung.bulkandbatchapi;
+package com.baeldung.bulkandbatchapi.request;
 
-import com.baeldung.bulkandbatchapi.customer.Customer;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
