+++ spring-web-modules/spring-rest-http-3/src/main/java/com/baeldung/bulkandbatchapi/response/BatchStatus.java
@@ -1,4 +1,4 @@
-package com.baeldung.bulkandbatchapi;
+package com.baeldung.bulkandbatchapi.response;
 
 public enum BatchStatus {
     PROCESSED, PARTIALLY_PROCESSED, NOT_PROCESSED
